feat: Templated-based updating of files
Extends the functionality of this library. Now it can take different types of repositories, and
frankly also add standardized repository files (not just build config). Consider renaming the
function?

BREAKING CHANGE: No longer, the templates are directly under templates/, but in each own folder of
repository types. All repositories require "service-config.yml" file to function.
