+++ Makefile
@@ -1,3 +1,3 @@
 .PHONY: build
 build:
-	GOOS=linux GOARCH=amd64 go build -o dist/update-repos main.go
+	GOOS=linux GOARCH=amd64 go build -o dist/update-repos cmd/devops-build-tools/main.go
