+++ go.mod
@@ -5,7 +5,9 @@ go 1.9
 require (
 	github.com/go-git/go-git/v5 v5.1.0
 	github.com/google/go-github/v32 v32.0.0
+	github.com/karrick/godirwalk v1.15.6
 	github.com/otiai10/copy v1.2.0
+	github.com/sabhiram/go-gitignore v0.0.0-20180611051255-d3107576ba94
 	golang.org/x/oauth2 v0.0.0-20200107190931-bf48bf16ab8d
 	gopkg.in/yaml.v2 v2.2.4
 )
