--- templates/.github/workflows/push.yml
@@ -1,69 +1,0 @@
-name: Release
-on:
-  push:
-    branches:
-      - master
-      - develop
-      - next
-      - alpha
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    if: "!contains(github.event.head_commit.message, 'ci skip')"
-    strategy:
-      matrix:
-        node-version: [12.x]
-    steps:
-    - uses: cobraz/branch-environment-action@master
-      with:
-        name: VAULT_ADDR
-      env:
-        DEVELOP_VAULT_ADDR: ${{ secrets.DEVELOP_VAULT_ADDR }}
-        MASTER_VAULT_ADDR: ${{ secrets.MASTER_VAULT_ADDR }}
-
-    - uses: bjerkio/vault-action@v1.1.3
-      id: vault
-      with:
-        endpoint: ${{ env.VAULT_ADDR }}
-        path: 'kv/data/devops/gcr-service-account'
-        authMethod: 'github'
-        githubToken: ${{ secrets.PERSONAL_GITHUB_TOKEN }}
-
-    - uses: actions/checkout@v2
-
-    - name: Semantic Release
-      id: semantic
-      uses: cycjimmy/semantic-release-action@v2
-      with:
-        semantic_version: 17.0.7
-        extra_plugins: |
-          @semantic-release/changelog@5.0.1
-          @semantic-release/git
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-  
-    # - uses: GoogleCloudPlatform/github-actions/setup-gcloud@master
-    #   with:
-    #     version: '290.0.1'
-    #     service_account_key: ${{ steps.vault.outputs.privateKey }}
-    #     export_default_credentials: true
-
-    # - run: gcloud --quiet auth configure-docker
-
-    - uses: docker://cobraz/pulumi-actions-bugfixing
-      with:
-        args: up
-      env:
-        PULUMI_ACCESS_TOKEN: ${{ secrets.PULUMI_ACCESS_TOKEN }}
-        GOOGLE_CREDENTIALS: ${{ steps.vault.outputs.privateKey }}
-        VERSION: ${{ steps.semantic.outputs.new_release_version }}
-        PULUMI_CI: up
-        PULUMI_ROOT: infra
-
-    - name: Request documentation update
-      uses: peter-evans/repository-dispatch@v1
-      with:
-          token: ${{ secrets.PERSONAL_GITHUB_TOKEN }}
-          repository: tabetalt/docs
-          event-type: update-docs
